Finish removal of RuntimeUtilities
